move buttons from calculate to select
